backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
pydantic>=2.4.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0
groq>=0.4.0
pillow>=10.1.0
numpy>=1.26.0
opencv-python>=4.8.1.78
langchain>=0.0.335
langchain-groq>=0.1.1
sounddevice>=0.4.6
soundfile>=0.12.1
speechrecognition>=3.10.0 